requirement.txt
Django==4.1
django-ckeditor==6.5.1
django-debug-toolbar==3.7.0
django-simple-captcha==0.5.17

